+++ packages/vue/src/range-slider/range-slider-context.ts
@@ -1,9 +1,7 @@
-import { type connect } from '@zag-js/range-slider'
-import { type ComputedRef } from 'vue'
 import { createContext } from '../context'
 import { type UseRangeSliderReturn } from './use-range-slider'
 
-export const [RangeSliderProvider, useRangeSliderContext] =
-  createContext<ComputedRef<ReturnType<typeof connect>>>('RangeSliderContext')
-
 export type RangeSliderContext = UseRangeSliderReturn
+
+export const [RangeSliderProvider, useRangeSliderContext] =
+  createContext<RangeSliderContext>('RangeSliderContext')