+++ packages/vue/src/range-slider/range-slider-marker-group.tsx
@@ -1,20 +1,19 @@
 import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import { getValidChildren, type ComponentWithProps } from '../utils'
+import { getValidChildren } from '../utils'
 import { useRangeSliderContext } from './range-slider-context'
 
 export type RangeSliderMarkerGroupProps = HTMLArkProps<'div'>
 
-export const RangeSliderMarkerGroup: ComponentWithProps<RangeSliderMarkerGroupProps> =
-  defineComponent({
-    name: 'RangeSliderMarkerGroup',
-    setup(_, { slots, attrs }) {
-      const api = useRangeSliderContext()
+export const RangeSliderMarkerGroup = defineComponent({
+  name: 'RangeSliderMarkerGroup',
+  setup(_, { slots, attrs }) {
+    const api = useRangeSliderContext()
 
-      return () => (
-        <ark.div {...api.value.markerGroupProps} {...attrs}>
-          {() => getValidChildren(slots)}
-        </ark.div>
-      )
-    },
-  })
+    return () => (
+      <ark.div {...api.value.markerGroupProps} {...attrs}>
+        {() => getValidChildren(slots)}
+      </ark.div>
+    )
+  },
+})