+++ packages/vue/src/range-slider/range-slider-track.tsx
@@ -1,11 +1,11 @@
 import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import { getValidChildren, type ComponentWithProps } from '../utils'
+import { getValidChildren } from '../utils'
 import { useRangeSliderContext } from './range-slider-context'
 
 export type RangeSliderTrackProps = HTMLArkProps<'div'>
 
-export const RangeSliderTrack: ComponentWithProps<RangeSliderTrackProps> = defineComponent({
+export const RangeSliderTrack = defineComponent({
   name: 'RangeSliderTrack',
   setup(_, { slots, attrs }) {
     const api = useRangeSliderContext()